Massive update, see the CHANGELOG (0.5.8) for more details
* Hashmarks: use RDF storage (old hashmarks db is deprecated)
* js-ipfs-http-client: upgrade to v49.0.2
* browser: Monkeypatch fetch() to support loading IPFS objects natively
  without any JS requirements
* Interceptor: add http://domain.eth to ens://domain.eth redirection
* Add a resource blocking system (with python-adblock; block lists are
  pulled from IPFS)
* Add the ips:// URL scheme (renders IPS JSON-LD schemas)
* Add a UI action to view IPS schemas
* Implement intelligent RDF tags (inspired from "Meaning of a tag")
* pronto: support for subconjunctive graphs (parent store)
* pubsub sniffer: nicer UI (add topic filter and max messages widgets)
* Add UI elements in the settings to configure the various webprofiles
* Add support for Greasemonkey scripts

revbump to 0.5.8

[deploy]

+++ galacteek/ui/resource.py
@@ -3,6 +3,7 @@ import os.path
 import asyncio
 import aioipfs
 
+from urllib.parse import unquote
 from typing import Union
 
 from rdflib import URIRef
@@ -109,7 +110,7 @@ class IPFSResourceOpener(QObject):
         webProfile = hashmark.get('prefsWebProfile', None)
 
         await self.open(
-            hashmark['uri'],  # URIRef
+            unquote(hashmark['uri']),  # URIRef, unquoted
             useWorkspace=workspace,
             schemePreferred=str(sPreferred) if sPreferred else None,
             minWebProfile=str(webProfile) if webProfile else None
@@ -122,7 +123,7 @@ class IPFSResourceOpener(QObject):
                    rdfGraph=None,
                    openingFrom=None,
                    pyramidOrigin=None,
-                   minWebProfile='ipfs',
+                   minWebProfile='minimal',
                    schemePreferred=None,
                    tryDecrypt=False,
                    fromEncrypted=False,
@@ -205,8 +206,7 @@ class IPFSResourceOpener(QObject):
                 wsSwitch = False
 
         if mimeType and mimeType.valid:
-            logUser.info('{path} ({type}): opening'.format(
-                path=rscPath, type=str(mimeType)))
+            log.debug(f'{rscPath} ({mimeType}): opening')
         else:
             logUser.info(iResourceCannotOpen(rscPath))
             return
@@ -489,7 +489,8 @@ class IPFSResourceOpener(QObject):
             os.unlink(filePath)
             log.debug(str(err))
         else:
-            os.unlink(filePath)
+            if os.path.isfile(filePath):
+                os.unlink(filePath)
 
     async def openWithSystemDefault(self, rscPath):
         # Use xdg-open or open depending on the platform
